feat: Unify CLI tools into a single file.
