+++ .eslintrc.js
@@ -11,5 +11,16 @@ module.exports = {
     "no-console": 0,
     "prefer-const": 2,
     "no-var": 2
-  }
+  },
+  overrides: [
+    {
+      files: ["bin/big.js"],
+      env: {
+        node: true
+      },
+      rules: {
+        "no-unused-expressions": "off"
+      }
+    }
+  ]
 };
