--- bin/template.hbs
@@ -1,19 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <meta charset='utf-8'>
-  <meta name='viewport' content='width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=0' />
-  <title>{{{ title }}}</title>
-  <link href='big.css' rel='stylesheet' type='text/css' />
-  <link href='highlight.css' rel='stylesheet' type='text/css' />
-  <style>
-    .new-shiny { background: #aaaaaa; }
-  </style>
-  <script src='big.js'></script>
-  <script src='highlight.js'></script>
-  <script>hljs.initHighlightingOnLoad();</script>
-</head>
-<body class='{{{ theme }}}'>
-  {{{ slides }}}
-</body>
-</html>
